+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/edit/abstracts/EditTextChatMessage.kt
@@ -3,5 +3,5 @@ package dev.inmo.tgbotapi.requests.edit.abstracts
 import dev.inmo.tgbotapi.abstracts.TextedOutput
 
 interface EditTextChatMessage : TextedOutput {
-    override val text: String
+    override val text: String?
 }